+++ include/sensor_buffer_struct.h
@@ -4,7 +4,7 @@
 #define LEN_RING_BUFF 10
 
 typedef struct {
-    float values[LEN_RING_BUFF];
+    double values[LEN_RING_BUFF];
     int write_index;
     int count;
 } SensorBuffer;
